Remove EmulatorValidation static class.
